\name{dgigamma}
\alias{dgigamma}
\title{generalized inverse gamma density}
\usage{
  dgigamma(x, alpha, beta, gamma = 1, mu = 0)
}
\arguments{
  \item{x}{vector of quantiles}

  \item{alpha}{alpha parameter}

  \item{beta}{beta parameter}

  \item{gamma}{gamma parameter}

  \item{mu}{mu parameter}
}
\description{
  generalized inverse gamma density
}

