\name{pgigamma}
\alias{pgigamma}
\title{distribution function for generalized inverse gamma}
\usage{
  pgigamma(q, alpha, beta, gamma = 1, mu = 0)
}
\arguments{
  \item{q}{vector of quantiles}

  \item{alpha}{alpha parameter}

  \item{beta}{beta parameter}

  \item{gamma}{gamma parameter}

  \item{mu}{mu parameter}
}
\description{
  distribution function for generalized inverse gamma
}

