\name{qgigamma}
\alias{qgigamma}
\title{quantile function for generalized inverse gamma}
\usage{
  qgigamma(p, alpha, beta, gamma = 1, mu = 0)
}
\arguments{
  \item{p}{vector of probabilities}

  \item{alpha}{alpha parameter}

  \item{beta}{beta parameter}

  \item{gamma}{gamma parameter}

  \item{mu}{mu parameter}
}
\description{
  quantile function for generalized inverse gamma
}

